Der-densor
Der-densor is mostly up and running now. It aborts if the densor is not 1-dimensional. I have tried it for B2 and B3 Lie algebras with success.

With this change I made a change in src/isom-test.m in PseudoIsometryGroup. There is no more ExpentiateDerivations, so it was replaced with IsometryGroup.